+++ src/helpers.ts
@@ -10,10 +10,17 @@ export const isStringLiteral = (
 export const defaultUrlProvider = (iconNamesParam: string) =>
   `https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0&${iconNamesParam}`;
 
-export const makeSelector = (jsxNodeRegex: RegExp, component: string) => {
+export const makeSelector = (
+  jsxNodeRegex: RegExp,
+  component: RegExp | string,
+) => {
   const calleeName = `/${jsxNodeRegex.source}/${jsxNodeRegex.flags}`;
+  const argName =
+    typeof component === "string"
+      ? `'${component}'`
+      : `/${component.source}/${component.flags}`;
   return [
-    `CallExpression[callee.name=${calleeName}][arguments.0.name='${component}']`,
+    `CallExpression[callee.name=${calleeName}][arguments.0.name=${argName}]`,
     ".arguments",
     "Property[key.name='children'] Literal",
   ].join(" > ");